+++ resources/publish.rb
@@ -20,21 +20,22 @@ property :publish_name,  String, name_property: true
 property :type,          String, default: ''
 property :endpoint,      String, default: ''
 property :component,     Array, default: []
-property :architectures, Array, default: ['amd64']
+property :architectures, Array, default: []
 property :prefix,        String, default: ''
 property :distribution,  String, default: ''
+property :timeout,       Integer, default: 3600
 
 action :create do
   components = new_resource.component.join(',')
-  architectures = new_resource.architectures.join(',')
   endpoint = new_resource.endpoint.empty? ? '' : "#{new_resource.endpoint}:"
 
   execute "Publish #{new_resource.type} - #{new_resource.publish_name}" do
-    command "aptly publish #{new_resource.type} -batch -passphrase='#{node['aptly']['gpg']['passphrase']}' -component='#{components}' -architectures='#{architectures}' -distribution='#{new_resource.distribution}' -- #{new_resource.publish_name} #{endpoint}#{new_resource.prefix}"
+    command "aptly publish #{new_resource.type} -batch -passphrase='#{node['aptly']['gpg']['passphrase']}' -component='#{components}' #{architectures(new_resource.architectures)} -distribution='#{new_resource.distribution}' -- #{new_resource.publish_name} #{endpoint}#{new_resource.prefix}"
     user node['aptly']['user']
     group node['aptly']['group']
     environment aptly_env
     sensitive true
+    timeout new_resource.timeout
     not_if %(aptly publish list | grep #{new_resource.publish_name})
   end
 end
@@ -46,6 +47,7 @@ action :update do
     user node['aptly']['user']
     group node['aptly']['group']
     environment aptly_env
+    timeout new_resource.timeout
   end
 end
 
@@ -58,6 +60,7 @@ action :drop do
     user node['aptly']['user']
     group node['aptly']['group']
     environment aptly_env
+    timeout new_resource.timeout
     only_if %(aptly publish list | grep #{new_resource.publish_name})
   end
 end